Created step def files
SD_Homepage
SD_Insights
SD_Common
AbstractPageStepDefinition
