+++ src/test/java/stepDefinitions/SD_Insights.java
@@ -2,6 +2,7 @@ package stepDefinitions;
 
 import org.openqa.selenium.WebDriver;
 
+import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
 import pageObjects.InsightsPage;
 
@@ -16,5 +17,22 @@ public class SD_Insights extends AbstractPageStepDefinition {
 		insightsPage.clickClearInsightKeywordBox();
 		insightsPage.insight_KeywordEnterText("Paul");	
 	}
+	
+	@When("^filtered Insights articles appear$")
+	public void filtered_Insights_articles_appear() throws Throwable {
+		
+	}
+
+	@When("^clicks Read More on Insights article$")
+	public void clicks_Read_More_on_Insights_article() throws Throwable {
+		System.out.println("Hello click on Read More");
+	}
+
+	@Then("^Insight article detail page opens$")
+	public void insight_article_detail_page_opens() throws Throwable {
+		System.out.println("Hello Articles detail page");
+		//driver.quit();
+	}
+	
 
 }
